html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-meta-refresh.optional.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: meta refresh</title>
<!-- This test is optional because the spec allows delaying or doing nothing for meta refresh. -->
<iframe src="support/blank.htm?src" loading="lazy"></iframe>
<script>
const iframe = document.querySelector('iframe');
const meta = iframe.contentDocument.createElement('meta');
meta.httpEquiv = 'Refresh';
meta.content = '0; url=support/blank.htm?nav';
iframe.contentDocument.head.append(meta);
</script>
<!-- Loading testharness.js here is intentional. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?nav", location.href).href);
    done();
  }, 250);
};
</script>
