LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-meta-refresh.optional-expected.txt



FAIL Navigating iframe loading='lazy' before it is loaded: meta refresh Error: assert_equals: expected "http://web-platform.test:8800/html/semantics/embedded-content/the-iframe-element/support/blank.htm?nav" but got "http://web-platform.test:8800/html/semantics/embedded-content/the-iframe-element/support/blank.htm?src"
